+++ fatal.cpp
@@ -24,9 +24,10 @@
 #include "unhandled.h"
 #include <stdlib.h>
 #include "cmdlineoptions.h"
+#include <cstdint>
 
 extern bool is_client;
-extern void message(const char*, const char*, unsigned long timeout_ms);
+extern void message(const char*, const char*, uint32_t timeout_ms);
 
 void fatal(const char* error, const char* extra)
 	{
